docs:remove main & all dir
